+++ projects/ks89/angular-modal-gallery/src/lib/components/plain-gallery/plain-gallery.component.ts
@@ -122,7 +122,7 @@ export class PlainGalleryComponent extends AccessibleComponent implements OnInit
   /**
    * Method ´ngOnInit´ to init both `configPlainGallery` calling `initPlainGalleryConfig()`
    * and `imageGrid invoking `initImageGrid()`.
-   * This is an Angular's lifecycle hook, so its called automatically by Angular itself.
+   * This is an angular lifecycle hook, so its called automatically by Angular itself.
    * In particular, it's called only one time!!!
    */
   ngOnInit(): void {
@@ -142,7 +142,7 @@ export class PlainGalleryComponent extends AccessibleComponent implements OnInit
 
   /**
    * Method ´ngOnChanges´ to update both `imageGrid` and`plainGalleryConfig`.
-   * This is an Angular's lifecycle hook, so its called automatically by Angular itself.
+   * This is an angular lifecycle hook, so its called automatically by Angular itself.
    * In particular, it's called when any data-bound property of a directive changes!!!
    */
   ngOnChanges(changes: SimpleChanges): void {