+++ projects/ks89/angular-modal-gallery/src/lib/components/upper-buttons/upper-buttons.component.ts
@@ -147,7 +147,7 @@ export class UpperButtonsComponent extends AccessibleComponent implements OnInit
   /**
    * Method ´ngOnInit´ to build `configButtons` applying a default value and also to
    * init the `buttons` array.
-   * This is an Angular's lifecycle hook, so its called automatically by Angular itself.
+   * This is an angular lifecycle hook, so its called automatically by Angular itself.
    * In particular, it's called only one time!!!
    */
   ngOnInit(): void {