+++ projects/ks89/angular-modal-gallery/src/lib/directives/direction.directive.ts
@@ -46,7 +46,7 @@ export class DirectionDirective implements OnInit, OnChanges {
 
   /**
    * Method ´ngOnInit´ to apply the style of this directive.
-   * This is an Angular's lifecycle hook, so its called automatically by Angular itself.
+   * This is an angular lifecycle hook, so its called automatically by Angular itself.
    * In particular, it's called only one time!!!
    */
   ngOnInit(): void {
@@ -55,7 +55,7 @@ export class DirectionDirective implements OnInit, OnChanges {
 
   /**
    * Method ´ngOnChanges´ to apply the style of this directive.
-   * This is an Angular's lifecycle hook, so its called automatically by Angular itself.
+   * This is an angular lifecycle hook, so its called automatically by Angular itself.
    * In particular, it's called when any data-bound property of a directive changes!!!
    */
   ngOnChanges(): void {